+++ src/legacy/core_plugins/data/public/search/expressions/build_tabular_inspector_data.ts
@@ -19,7 +19,7 @@
 
 import { set } from 'lodash';
 import { FormattedData } from '../../../../../../plugins/inspector/public';
-import { TabbedTable } from '../tabify';
+import { TabbedTable } from '../../../../../../plugins/data/public';
 import { createFilter } from './create_filter';
 
 /**
