+++ src/legacy/ui/public/agg_response/index.js
@@ -19,10 +19,10 @@
 
 import { buildHierarchicalData } from './hierarchical/build_hierarchical_data';
 import { buildPointSeriesData } from './point_series/point_series';
-import { tabifyAggResponse } from '../../../core_plugins/data/public';
+import { search } from '../../../../plugins/data/public';
 
 export const aggResponseIndex = {
   hierarchical: buildHierarchicalData,
   pointSeries: buildPointSeriesData,
-  tabify: tabifyAggResponse,
+  tabify: search.tabifyAggResponse,
 };
