+++ src/plugins/data/public/search/tabify/get_columns.test.ts
@@ -19,13 +19,8 @@
 
 import { tabifyGetColumns } from './get_columns';
 import { TabbedAggColumn } from './types';
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { AggConfigs } from '../../../../../../plugins/data/public/search/aggs';
-import {
-  mockAggTypesRegistry,
-  mockDataServices,
-  // eslint-disable-next-line @kbn/eslint/no-restricted-paths
-} from '../../../../../../plugins/data/public/search/aggs/test_helpers';
+import { AggConfigs } from '../aggs';
+import { mockAggTypesRegistry, mockDataServices } from '../aggs/test_helpers';
 
 describe('get columns', () => {
   beforeEach(() => {
