+++ src/plugins/data/public/search/tabify/tabify.test.ts
@@ -18,11 +18,9 @@
  */
 
 import { tabifyAggResponse } from './tabify';
-import { IndexPattern, IAggConfig, IAggConfigs } from '../../../../../../plugins/data/public';
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { AggConfigs } from '../../../../../../plugins/data/public/search/aggs';
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { mockAggTypesRegistry } from '../../../../../../plugins/data/public/search/aggs/test_helpers';
+import { IndexPattern } from '../../../public';
+import { AggConfigs, IAggConfig, IAggConfigs } from '../aggs';
+import { mockAggTypesRegistry } from '../aggs/test_helpers';
 import { metricOnly, threeTermBuckets } from 'fixtures/fake_hierarchical_data';
 
 describe('tabifyAggResponse Integration', () => {
