+++ src/plugins/data/public/search/tabify/tabify.ts
@@ -21,10 +21,8 @@ import { get } from 'lodash';
 import { TabbedAggResponseWriter } from './response_writer';
 import { TabifyBuckets } from './buckets';
 import { TabbedResponseWriterOptions, TabbedRangeFilterParams } from './types';
-import { AggResponseBucket } from '../types';
-import { IAggConfigs } from '../../../../../../plugins/data/public';
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { AggGroupNames } from '../../../../../../plugins/data/public/search/aggs';
+import { AggResponseBucket } from './types';
+import { AggGroupNames, IAggConfigs } from '../aggs';
 
 /**
  * Sets up the ResponseWriter and kicks off bucket collection.
